+++ plone/app/caching/tests/test_lookup.py
@@ -1,5 +1,5 @@
 import unittest2 as unittest
-from plone.testing.zca import UNIT_TESTING
+from plone.caching.testing import IMPLICIT_RULESET_REGISTRY_UNIT_TESTING
 
 from z3c.caching.registry import RulesetRegistry
 import z3c.caching.registry
@@ -78,11 +78,10 @@ class DummyView(BrowserView):
 
 class TestContentItemLookup(unittest.TestCase):
 
-    layer = UNIT_TESTING
+    layer = IMPLICIT_RULESET_REGISTRY_UNIT_TESTING
 
     def setUp(self):
         provideAdapter(persistentFieldAdapter)
-        provideAdapter(RulesetRegistry)
 
     def test_no_registry(self):
         published = ZopePageTemplate('someView')